Move API router handlers into their own files
